+++ __tests__/clients/inworld.client.spec.ts
@@ -9,6 +9,7 @@ import {
   extension,
   generateSessionToken,
   phrases,
+  previousState,
   SCENE,
   user,
 } from '../helpers';
@@ -47,7 +48,7 @@ describe('should finish with success', () => {
       .setOnReady(onReady)
       .setOnHistoryChange(onHistoryChange)
       .setOnPhoneme(onPhoneme)
-      .setSessionContinuation({ previousDialog: phrases });
+      .setSessionContinuation({ previousState, previousDialog: phrases });
   });
 
   test('should open connection', async () => {