+++ __tests__/helpers.ts
@@ -143,3 +143,4 @@ export const phrases = [
   },
 ];
 export const previousDialog = new PreviousDialog(phrases);
+export const previousState = new Uint8Array(32);